perf: code-split the dev-only DiagnosticsPanel out of the production bundle
